refactor: widget-data type 위치 이동
