+++ app/(home)/(default-home)/@widgetSection/layout.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const WidgetSectionLayout = ({ children }: { children: React.ReactNode }) => (
-  <section className='px-20'>{children}</section>
+  <section className='mb-24 px-20'>{children}</section>
 );
 
 export default WidgetSectionLayout;
