+++ app/(home)/(default-home)/@widgetSection/page.tsx
@@ -52,7 +52,7 @@ const WidgeSectionPage = async () => {
       <ConsumeWeatherCard />
 
       {/* 위젯 영역 */}
-      <div className='mb-[3rem] grid grid-cols-2 justify-items-center gap-[1.6rem]'>
+      <div className='grid grid-cols-2 justify-items-center gap-[1.6rem]'>
         {data[0].showWidget.map((item) => {
           const WidgetComponent = widgetMap[item.id];
           return WidgetComponent ? <WidgetComponent key={item.id} /> : null;
