+++ app/(home)/edit-widget/types/index.ts
@@ -1,14 +1,5 @@
 import { UniqueIdentifier } from '@dnd-kit/core/dist/types';
-export type EditWidgetDataType = {
-  showWidget: {
-    id: UniqueIdentifier;
-    title: string;
-  }[];
-  hideWidget: {
-    id: UniqueIdentifier;
-    title: string;
-  }[];
-}[];
+
 export type SetHideWidgetType = React.Dispatch<
   React.SetStateAction<
     {
