+++ service/types/widget-type/index.ts
@@ -1,0 +1,11 @@
+import { UniqueIdentifier } from '@dnd-kit/core/dist/types';
+export type EditWidgetDataType = {
+  showWidget: {
+    id: UniqueIdentifier;
+    title: string;
+  }[];
+  hideWidget: {
+    id: UniqueIdentifier;
+    title: string;
+  }[];
+}[];
